Split tests in test_clib.py into mutliple smaller test files
